passes-pdf-ui: remove inline pinch-zoom from DocumentView (wpass-ny4)
Per the wpass-6ag pivot, inline DocumentView returns to a fixed 1x
display: no pinch-zoom, no pan, no double-tap. The gesture surface
introduced in wpass-1wq is extracted to internal/ZoomableImage.kt so the
full-screen detail surface (wpass-jil) is the only call site.

Instrumented tests updated:
- pinchOnTheInlineSurfaceDoesNotZoomAndKeepsTheTrustCaptionVisible
  replaces the inline pinchToZoom assertion as a regression for the
  removed inline-zoom behaviour.
- singleTouchHorizontalDragAtFitScaleStillAdvancesThePager kept as the
  swipe-to-page contract.
- inline-zoom-specific cases (doubleTap, drag-while-zoomed) removed;
  those behaviours now live on the full-screen surface.

ADR 0005 D5 trust caption behaviour inline is unchanged.
